# Improve performance of status map parser