--- a/go.mod
+++ b/go.mod
@@ -1,5 +1,3 @@
 module github.com/mackerelio/checkers
 
-go 1.18
-
-require golang.org/x/exp v0.0.0-20230905200255-921286631fa9
+go 1.20