This PR proposes improvement of #20. The new implementation reduces allocations as possible, and about 30% faster. Also this way of parsing is more Go-way.

```
func BenchmarkParse(b *testing.B) {
	for i := 0; i < b.N; i++ {
		Parse("unknown=critical,warning=critical")
	}
}

name     old time/op    new time/op    delta
Parse-8     429ns ± 0%     296ns ± 1%  -31.06%  (p=0.000 n=9+10)

name     old alloc/op   new alloc/op   delta
Parse-8      256B ± 0%      240B ± 0%   -6.25%  (p=0.000 n=10+10)

name     old allocs/op  new allocs/op  delta
Parse-8      7.00 ± 0%      3.00 ± 0%  -57.14%  (p=0.000 n=10+10)
```