@lufia 2023-10-02
@itchyny thanks!

... CIs are still broken because Coveralls rejects the uploads when the coverage uses same build-id to any one of past builds.

Unfortunately we cannot delete the build from Coveralls, so I'm going to merge this to base branch, then runs CIs in that.